// web/src/lib/api/generated/models/poskasir-internal-orders-repository-order-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const POSKasirInternalOrdersRepositoryOrderStatus = {
    OrderStatusOpen: 'open',
    OrderStatusInProgress: 'in_progress',
    OrderStatusServed: 'served',
    OrderStatusPaid: 'paid',
    OrderStatusCancelled: 'cancelled'
} as const;

export type POSKasirInternalOrdersRepositoryOrderStatus = typeof POSKasirInternalOrdersRepositoryOrderStatus[keyof typeof POSKasirInternalOrdersRepositoryOrderStatus];